profile now, render independent user components and prop-type has been addded
