+++ client/package.json
@@ -44,6 +44,5 @@
   "devDependencies": {
     "faker": "^4.1.0",
     "fs": "^0.0.1-security"
-  },
-//  "proxy": "https://jobmatcher-api-stage.herokuapp.com"
+  }
 }
